auth_service/requirements.txt
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# Authentication
PyJWT>=2.8.0
cryptography>=42.0.0

# Utilities
python-multipart>=0.0.6
